Move resolveEntityAndEndpoint() function to zigbee.ts
